Moved model classes to separate folder
