+++ Tests/Schema.NET.Test/ValuesJsonConverterTest.cs
@@ -547,42 +547,3 @@ namespace Schema.NET.Test
         }
     }
 }
-
-#pragma warning disable SA1649 // File name should match first type name
-#pragma warning disable SA1403 // File may only contain a single namespace
-#pragma warning disable SA1402 // File may only contain a single type
-namespace SomeCustomNamespace
-{
-    using System.Runtime.Serialization;
-    using Newtonsoft.Json;
-    using Schema.NET;
-
-    public class ExternalSchemaModelCustomNamespace : Thing
-    {
-        [DataMember(Name = "@type")]
-        public override string Type => "ExternalSchemaModelCustomNamespace";
-
-        [DataMember(Name = "myCustomProperty")]
-        [JsonConverter(typeof(ValuesJsonConverter))]
-        public OneOrMany<string> MyCustomProperty { get; set; }
-    }
-}
-
-namespace Schema.NET
-{
-    using System.Runtime.Serialization;
-    using Newtonsoft.Json;
-
-    public class ExternalSchemaModelSharedNamespace : Thing
-    {
-        [DataMember(Name = "@type")]
-        public override string Type => "ExternalSchemaModelSharedNamespace";
-
-        [DataMember(Name = "myCustomProperty")]
-        [JsonConverter(typeof(ValuesJsonConverter))]
-        public OneOrMany<string> MyCustomProperty { get; set; }
-    }
-}
-#pragma warning restore SA1402 // File may only contain a single type
-#pragma warning restore SA1403 // File may only contain a single namespace
-#pragma warning restore SA1649 // File name should match first type name
